+++ src/CastTo/PascalCase.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Nandan108\DtoToolkit\CastTo;
+
+use Nandan108\DtoToolkit\Core\BaseDto;
+use Nandan108\DtoToolkit\Core\CastBase;
+use Nandan108\DtoToolkit\Support\CaseConverter;
+
+#[\Attribute(\Attribute::TARGET_PROPERTY | \Attribute::IS_REPEATABLE)]
+final class PascalCase extends CastBase
+{
+    #[\Override]
+    public function cast(mixed $value, array $args, BaseDto $dto): string
+    {
+        return CaseConverter::toPascal($this->throwIfNotStringable($value));
+    }
+}